Move upload button to Footer and uploadChallenge to App
